FEATURE: Extract DimensionSelector and render single dimentsion
This commit extracts the DimensionSelector to an own component. The component gets an additional property to render only the Select without any label.

The DimensionSwitcher now renders a single DimensionSelector if we have just one dimension and the old dropdown if we have multiple.
